Move numtracker from client package into utils
The numtracker is not used by the client.
