+++ src/blueapi/utils/__init__.py
@@ -3,6 +3,7 @@ from .connect_devices import connect_devices
 from .file_permissions import get_owner_gid, is_sgid_set
 from .invalid_config_error import InvalidConfigError
 from .modules import is_function_sourced_from_module, load_module_all
+from .numtracker import NumtrackerClient
 from .serialization import serialize
 from .thread_exception import handle_all_exceptions
 
@@ -14,6 +15,7 @@ __all__ = [
     "BlueapiModelConfig",
     "BlueapiPlanModelConfig",
     "InvalidConfigError",
+    "NumtrackerClient",
     "connect_devices",
     "is_sgid_set",
     "get_owner_gid",
